requirements.txt
torch>=2.3.0
transformers>=4.45.0
tokenizers>=0.20.0
accelerate>=0.33.0
peft>=0.12.0
datasets>=2.20.0
pillow>=10.4.0
numpy>=1.26.0
tqdm>=4.66.0
requests>=2.32.0
beautifulsoup4>=4.12.3
wandb>=0.17.0
einops>=0.8.0
addict>=2.4.0
easydict>=1.13

